learn_django/requirements.txt
aioredis==1.3.1
ansys-dpf-core==0.4.1
ansys-grpc-dpf==0.4.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
bcrypt==3.2.0
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Django==4.0.4
django-plotly-dash==1.6.6
dpd-components==0.1.0
Flask==2.1.1
Flask-Compress==1.11
future==0.18.2
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.22.3
pandas==1.4.2
paramiko==2.10.3
plotly==5.7.0
progressbar2==4.0.0
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-utils==3.1.0
pytz==2022.1
requests==2.27.1
rsa==4.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
xmltodict==0.12.0
zipp==3.8.0
zope.interface==5.4.0
